# shared/api-schema/openapi.yaml
openapi: "3.0.3"
info:
  title: "Personality API"
  version: "1.0.0"
  description: >
    Personality portal + tarot mobile API.
    Phase 1 conversion — Hono Workers backend.
servers:
  - url: "https://api.personality.app"
    description: "Production"
  - url: "http://localhost:8787"
    description: "Local development (wrangler dev)"
paths:
...
  /admin/audit_logs:
    get:
      operationId: listAdminAuditLogs
      summary: List audit logs
      tags: [admin, audit_logs]
      security:
        - CfAccessJwt: []
      responses:
        "200":
          description: Audit log list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/audit_logs/{id}:
    get:
      operationId: getAdminAuditLog
      summary: Get single audit log
      tags: [admin, audit_logs]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin: Question Sets ─────────────────────────────────────────────────────
  /admin/question_sets:
    get:
      operationId: listAdminQuestionSets
      summary: List question sets
      tags: [admin, question_sets]
      security:
        - CfAccessJwt: []
      responses:
        "200":
          description: Question set list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    post:
      operationId: createAdminQuestionSet
      summary: Create question set
      tags: [admin, question_sets]
      security:
        - CfAccessJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                version:
                  type: string
      responses:
        "201":
          description: Question set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/question_sets/{id}:
    get:
      operationId: getAdminQuestionSet
      summary: Get question set
      tags: [admin, question_sets]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminQuestionSet
      summary: Update question set
      tags: [admin, question_sets]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Question set updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: deleteAdminQuestionSet
      summary: Delete question set
      tags: [admin, question_sets]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question set deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin: Alerts ────────────────────────────────────────────────────────────
  /admin/alerts:
    get:
      operationId: listAdminAlerts
      summary: List system alerts
      tags: [admin, alerts]
      security:
        - CfAccessJwt: []
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/alerts/{id}:
    get:
      operationId: getAdminAlert
      summary: Get single alert
      tags: [admin, alerts]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: acknowledgeAdminAlert
      summary: Acknowledge / update alert
      tags: [admin, alerts]
      security:
        - CfAccessJwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin: Dashboard ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CfAccessJwt:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
  schemas:
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: "Response payload (type varies by endpoint)"
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation_failed
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - internal_error
            message:
              type: string
            details:
              description: "Optional structured details"